chore: move keyring snap permissions file
