+++ app/scripts/lib/snap-keyring/keyring-snaps-permissions.test.ts
