+++ app/scripts/lib/snap-keyring/keyring-snaps-permissions.ts
